Fix `e` execution option
The bug was caused by the migration from a large query string that gets
sliced into statements afterwards by splitting by `;`, to the splitting
by the `sqlparse` module and returning a tuple of individual statements.
Also from having these statement providers as class `@property` members,
to just class methods. Since the `query` property was an actual class
field, it was skipped in previous refactoring processes. For
consistency, the `query` option was renamed to `execute_stmt`, and a
method was added in the base `Query` class.